Add doxygen documentation and rename files of generarting mesh data
